[Update] File

+++ src/file/manager.py
@@ -34,7 +34,13 @@ def check_path_existing(path: str):
 
 
 def check_file_type(path: str, fileType: str):
-    pass
+    if not fileType.startswith('.'):
+        fileType = '.'+fileType
+    return not os.path.isdir(path) and os.path.isfile(path) and path.endswith(fileType)
+
+
+def create_dir(path: str, exist_ok: bool = True):
+    os.makedirs(path, exist_ok=exist_ok)
 
 
 def remove_path(path: str):
@@ -45,8 +51,9 @@ def remove_file(path: str):
     os.remove(path)
 
 
-def remove_auto(path: str):
-    if os.path.isfile(path):
-        remove_file(path)
-    else:
-        remove_path(path)
+def remove_auto(path: str, non_exist_ok: bool = True):
+    if check_path_existing(path) or not non_exist_ok:
+        if os.path.isfile(path):
+            remove_file(path)
+        else:
+            remove_path(path)
